// src/git.d.ts
export interface Git {
  readonly path: string;
}
// ...
export interface Repository {
  readonly rootUri: Uri;
  readonly inputBox: InputBox;
  readonly state: RepositoryState;
  readonly ui: RepositoryUIState;

  getConfigs(): Promise<{ key: string; value: string }[]>;
  getConfig(key: string): Promise<string>;
  setConfig(key: string, value: string): Promise<string>;

  getObjectDetails(
    treeish: string,
    path: string
  ): Promise<{ mode: string; object: string; size: number }>;
  detectObjectType(
    object: string
  ): Promise<{ mimetype: string; encoding?: string }>;
  buffer(ref: string, path: string): Promise<Buffer>;
  show(ref: string, path: string): Promise<string>;
  getCommit(ref: string): Promise<Commit>;

  clean(paths: string[]): Promise<void>;

  diffWithHEAD(path: string): Promise<string>;
  diffWith(ref: string, path: string): Promise<string>;
  diffIndexWithHEAD(path: string): Promise<string>;
  diffIndexWith(ref: string, path: string): Promise<string>;
  diffBlobs(object1: string, object2: string): Promise<string>;
  diffBetween(ref1: string, ref2: string, path: string): Promise<string>;

  hashObject(data: string): Promise<string>;

  createBranch(name: string, checkout: boolean, ref?: string): Promise<void>;
  deleteBranch(name: string): Promise<void>;
  getBranch(name: string): Promise<Branch>;
  setBranchUpstream(name: string, upstream: string): Promise<void>;

  getMergeBase(ref1: string, ref2: string): Promise<string>;

  status(): Promise<void>;
  checkout(treeish: string): Promise<void>;

  addRemote(name: string, url: string): Promise<void>;
  removeRemote(name: string): Promise<void>;

  fetch(remote?: string, ref?: string): Promise<void>;
  pull(): Promise<void>;
}

export interface API {
  readonly git: Git;
  readonly repositories: Repository[];
  readonly onDidOpenRepository: Event<Repository>;
  readonly onDidCloseRepository: Event<Repository>;
}

export interface GitExtension {
  /**
   * Returns a specific API version.
   *
   * @param version Version number.
   * @returns API instance
   */
  getAPI(version: 1): API;
}